Android: change native pointer types from int to long in src/printing
BUG=317523

Review URL: https://codereview.chromium.org/83283002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@236820 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 879517bd47bd27255ae5d29a1b6cfe3b2bb793fe

+++ printing/android/java/src/org/chromium/printing/PrintingContext.java
@@ -32,9 +32,9 @@ public class PrintingContext implements PrintingContextInterface {
     private final PrintingController mController;
 
     /** The pointer to the native PrintingContextAndroid object. */
-    private final int mNativeObject;
+    private final long mNativeObject;
 
-    private PrintingContext(int ptr) {
+    private PrintingContext(long ptr) {
         mController = PrintingControllerImpl.getInstance();
         mNativeObject = ptr;
     }
@@ -129,6 +129,6 @@ public class PrintingContext implements PrintingContextInterface {
     }
 
     private native void nativeAskUserForSettingsReply(
-            int nativePrintingContextAndroid,
+            long nativePrintingContextAndroid,
             boolean success);
 }
